Create plsql_schema_cleaner.sql
This anonymous PL/SQL block drops all the objects of a given schema/user except for DB links. It's intended use is to clear a schema prior to an import. or testing purposes, the execute immediate lines are commented. You're free to try it and uncomment after you have checked that it will produce your desired results.